@@ -3,20 +3,22 @@
   "version": "0.2.0",
   "license": "MIT",
   "repository": "LinusU/load-yaml-file",
-  "engines": {
-    "node": ">=6"
-  },
+  "type": "module",
+  "exports": "./index.js",
   "scripts": {
-    "test": "standard && mocha"
+    "test": "standard && mocha && ts-readme-generator --check"
   },
   "dependencies": {
-    "graceful-fs": "^4.1.5",
+    "graceful-fs": "^4.2.8",
     "js-yaml": "^4.1.0",
-    "pify": "^4.0.1",
-    "strip-bom": "^3.0.0"
+    "strip-bom": "^5.0.0"
   },
   "devDependencies": {
-    "mocha": "^6.0.2",
-    "standard": "^12.0.1"
+    "mocha": "^9.1.0",
+    "standard": "^16.0.3",
+    "ts-readme-generator": "^0.6.1"
+  },
+  "engines": {
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
   }
 }
